upscale HRRR-ZARR
Hardwired for HRRRv4 and 80km grid.

rename DataArrays so they don't conflict. There are many called CAPE, so you need to
use the long_name attribute, which is the level.

Rename and upscale in the proprocess function. It runs in parallel if extra cpus are
available. For 50 variables, 5 cpus is optimal.

Work with units, and derive necessary fields with metpy.

Save Pandas DataFrame as parquet.