Added --ifile argument
renamed ifile0 ifile

Don't use generator for cvsplit . When defined and initialized before
the statsjob function, it will work as expected the first time statsjob
is run, but be exhausted by the time statsjob runs a 2nd time. Use a list
instead.

include the # of samples in the output

remove chunksize. Don't think it helped.