Had to revert to old version of Tony's ML model directories.
They were all in the nn/ subdirectory, which was ignored by git because
nn/ was in .gitignore. When I used git mv source target it kept the
old target directory name and just moved it to a subdirectory named
after source.

I had to remove nn/ from the .gitignore file before retrying the
git mv commands. I used a script called nn/fix_git_yaml_dirnames.csh.

There are a lot of ignored nn files now that nn is not in the .gitignore
file. That may need revisiting but is in line with the trend toward
keeping a snapshot of "everything".