initialize new optimizer with each model
was getting error about not finding the dense layer0 when the
second iteration of the fit loop was hit. I think you can't
initialize the optimizer outside the loop and reuse it in the loop.

Moved clear_session to end of loop. Not sure if this change or the
optimizer change allowed multiple fits to start working again in the
same job.